chore: move files around, add meta export, remove base export, add EvidenceError to util export
