--- packages/lib/sdk/src/index.js
@@ -1,5 +1,0 @@
-/// <reference path="./types/index.d.ts"/>
-export { evidencePlugin as evidenceVite } from './build-dev/vite/index.js';
-export { svelte as evidenceSvelte } from './build-dev/svelte/index.js';
-
-export { getEvidenceConfig } from './configuration/getEvidenceConfig.js';
