+++ packages/lib/sdk/src/lib/projectRoot.js
@@ -6,4 +6,7 @@ if (_init.endsWith(path.join('.evidence', 'template'))) {
 	_init = _init.substring(0, _init.length - path.join('.evidence', 'template').length);
 }
 
+/**
+ * @deprecated Use projectRoot from "projectPaths.js" instead
+ */
 export const projectRoot = _init;
