+++ packages/lib/sdk/src/meta.js
@@ -1,0 +1,1 @@
+export * as paths from './lib/projectPaths.js';
