+++ packages/lib/sdk/src/plugins/datasources/evalSources.js
@@ -10,7 +10,7 @@ import { buildMultipartParquet } from '@evidence-dev/universal-sql';
 import { buildSourceDirectoryProxy } from './buildSourceDirectoryProxy.js';
 import { addToCache, checkCache, flushCache, loadCache } from './SourceResultCache.js';
 import ora from 'ora';
-import { dataUrlPrefix } from '../../build-dev/vite/virtuals/node/projectPaths.js';
+import { dataUrlPrefix } from '../../lib/projectPaths.js';
 import { subSourceVariables } from './sub-source-vars.js';
 import { logQueryEvent } from '@evidence-dev/telemetry';
