+++ packages/lib/sdk/src/utils/index.js
@@ -1,2 +1,3 @@
+export { EvidenceError } from '../lib/EvidenceError.js';
 export * from './batchUp.js';
 export * from '../lib/sharedPromise.js';
